tryfix: heltec-wireless-tracker periodic config loss
The Heltec Wireless Tracker and its UC6580 doesn't support a soft
sleep, so we hard sleep it.

It was discovered that this causes the gnss chip to lose its settings.
This means the buffer overflows with NMEA messages we don't use, and
constellations that were previously enabled may be disabled.

This patch attempts to fix it by performing the configuration of the
UC6580 chip every time it comes back from hardsleep.

Fixes https://github.com/meshtastic/firmware/issues/5088